Fix flaky e2e tests (#2383)
* Test grpc flake and repro it
* Fix bug in test where we never cleanup envoy
* Fix transformation tests for mac local dev
* Add changelog
* Revert "Test grpc flake and repro it"

This reverts commit 6459db08f2109f829dead1a187495ad913ef34fd.

+++ test/e2e/consul_test.go
@@ -91,7 +91,7 @@ var _ = Describe("Consul e2e", func() {
 		testClients = services.RunGlooGatewayUdsFds(ctx, ro)
 
 		// Start Envoy
-		envoyPort = uint32(defaults.HttpPort)
+		envoyPort = defaults.HttpPort
 		envoyInstance, err = envoyFactory.NewEnvoyInstance()
 		Expect(err).NotTo(HaveOccurred())
 		err = envoyInstance.RunWithRole(writeNamespace+"~"+gatewaydefaults.GatewayProxyName, testClients.GlooPort)
